refactor: Move IoxMetadata to parquet_file
