+++ ingester/Cargo.toml
@@ -20,6 +20,7 @@ mutable_batch = { path = "../mutable_batch"}
 object_store = { path = "../object_store" }
 parking_lot = "0.11.2"
 parquet = "8.0"
+parquet_file = { path = "../parquet_file" }
 predicate = { path = "../predicate" }
 prost = "0.9"
 query = { path = "../query" }
