+++ ingester/src/persist.rs
@@ -1,116 +1,21 @@
 //! Persist compacted data to parquet files in object storage
 
-use arrow::{datatypes::SchemaRef, record_batch::RecordBatch};
+use arrow::record_batch::RecordBatch;
 use bytes::Bytes;
-use generated_types::influxdata::iox::ingest::v1 as proto;
-use iox_catalog::interface::{NamespaceId, PartitionId, SequenceNumber, SequencerId, TableId};
 use object_store::{
     path::{ObjectStorePath, Path},
     ObjectStore, ObjectStoreApi,
 };
-use parking_lot::Mutex;
-use parquet::{
-    arrow::ArrowWriter,
-    basic::Compression,
-    file::{metadata::KeyValue, properties::WriterProperties, writer::TryClone},
-};
-use prost::Message;
-use snafu::{OptionExt, ResultExt, Snafu};
-use std::{
-    io::{Cursor, Seek, SeekFrom, Write},
-    sync::Arc,
-};
-use time::Time;
-use uuid::Uuid;
-
-/// IOx-specific metadata.
-///
-/// # Serialization
-/// This will serialized as base64-encoded [Protocol Buffers 3] into the file-level key-value
-/// Parquet metadata (under [`METADATA_KEY`]).
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct IoxMetadata {
-    /// The uuid used as the location of the parquet file in the OS.
-    /// This uuid will later be used as the catalog's ParquetFileId
-    pub object_store_id: Uuid,
-
-    /// Timestamp when this file was created.
-    pub creation_timestamp: Time,
-
-    /// namespace id of the data
-    pub namespace_id: NamespaceId,
-
-    /// namespace name of the data
-    pub namespace: Arc<str>,
-
-    /// sequencer id of the data
-    pub sequencer_id: SequencerId,
-
-    /// table id of the data
-    pub table_id: TableId,
-
-    /// table name of the data
-    pub table_name: Arc<str>,
-
-    /// partition id of the data
-    pub partition_id: PartitionId,
-
-    /// parittion key of the data
-    pub partition_key: Arc<str>,
-
-    /// Time of the first write of the data
-    /// This is also the min value of the column `time`
-    pub time_of_first_write: Time,
-
-    /// Time of the last write of the data
-    /// This is also the max value of the column `time`
-    pub time_of_last_write: Time,
-
-    /// sequence number of the first write
-    pub min_sequence_number: SequenceNumber,
-
-    /// sequence number of the last write
-    pub max_sequence_number: SequenceNumber,
-}
-
-impl IoxMetadata {
-    /// Convert to protobuf v3 message.
-    fn to_protobuf(&self) -> std::result::Result<Vec<u8>, prost::EncodeError> {
-        let proto_msg = proto::IoxMetadata {
-            object_store_id: self.object_store_id.as_bytes().to_vec(),
-            creation_timestamp: Some(self.creation_timestamp.date_time().into()),
-            namespace_id: self.namespace_id.get(),
-            namespace: self.namespace.to_string(),
-            sequencer_id: self.sequencer_id.get() as i32,
-            table_id: self.table_id.get(),
-            table_name: self.table_name.to_string(),
-            partition_id: self.partition_id.get(),
-            partition_key: self.partition_key.to_string(),
-            time_of_first_write: Some(self.time_of_first_write.date_time().into()),
-            time_of_last_write: Some(self.time_of_last_write.date_time().into()),
-            min_sequence_number: self.min_sequence_number.get(),
-            max_sequence_number: self.max_sequence_number.get(),
-        };
-
-        let mut buf = Vec::new();
-        proto_msg.encode(&mut buf)?;
-
-        Ok(buf)
-    }
-}
-
-/// File-level metadata key to store the IOx-specific data.
-///
-/// This will contain [`IoxMetadata`] serialized as base64-encoded [Protocol Buffers 3].
-///
-/// [Protocol Buffers 3]: https://developers.google.com/protocol-buffers/docs/proto3
-pub const METADATA_KEY: &str = "IOX:metadata";
+use parquet_file::metadata::IoxMetadata;
+use snafu::{ResultExt, Snafu};
 
 #[derive(Debug, Snafu)]
 #[allow(missing_docs)]
 pub enum Error {
     #[snafu(display("Error converting the parquet stream to bytes: {}", source))]
-    ConvertingToBytes { source: ParquetBytesError },
+    ConvertingToBytes {
+        source: parquet_file::storage::Error,
+    },
 
     #[snafu(display("Error writing to object store: {}", source))]
     WritingToObjectStore { source: object_store::Error },
@@ -133,7 +38,7 @@ pub async fn persist(
         .expect("record_batches.is_empty was just checked")
         .schema();
 
-    let data = parquet_bytes(metadata, record_batches, schema)
+    let data = parquet_file::storage::Storage::parquet_bytes(record_batches, schema, metadata)
         .await
         .context(ConvertingToBytesSnafu)?;
 
@@ -168,112 +73,15 @@ fn parquet_file_object_store_path(metadata: &IoxMetadata, object_store: &ObjectS
     path
 }
 
-#[derive(Debug, Snafu)]
-#[allow(missing_docs)]
-pub enum ParquetBytesError {
-    #[snafu(display("Cannot encode metadata: {}", source))]
-    MetadataEncodeFailure { source: prost::EncodeError },
-
-    #[snafu(display("Error opening Parquet Writer: {}", source))]
-    OpeningParquetWriter {
-        source: parquet::errors::ParquetError,
-    },
-
-    #[snafu(display("Error writing Parquet to memory: {}", source))]
-    WritingParquetToMemory {
-        source: parquet::errors::ParquetError,
-    },
-
-    #[snafu(display("Error closing Parquet Writer: {}", source))]
-    ClosingParquetWriter {
-        source: parquet::errors::ParquetError,
-    },
-
-    #[snafu(display("Error converting to vec[u8]: Nothing else should have a reference here"))]
-    WritingToMemWriter {},
-}
-
-/// Convert the given metadata and RecordBatches to parquet file bytes
-async fn parquet_bytes(
-    metadata: &IoxMetadata,
-    record_batches: &[RecordBatch],
-    schema: SchemaRef,
-) -> Result<Vec<u8>, ParquetBytesError> {
-    let metadata_bytes = metadata.to_protobuf().context(MetadataEncodeFailureSnafu)?;
-    let props = writer_props(&metadata_bytes);
-
-    let mem_writer = MemWriter::default();
-    {
-        let mut writer = ArrowWriter::try_new(mem_writer.clone(), schema, Some(props))
-            .context(OpeningParquetWriterSnafu)?;
-        for batch in record_batches {
-            writer.write(batch).context(WritingParquetToMemorySnafu)?;
-        }
-        writer.close().context(ClosingParquetWriterSnafu)?;
-    } // drop the reference to the MemWriter that the SerializedFileWriter has
-
-    mem_writer.into_inner().context(WritingToMemWriterSnafu)
-}
-
-fn writer_props(metadata_bytes: &[u8]) -> WriterProperties {
-    WriterProperties::builder()
-        .set_key_value_metadata(Some(vec![KeyValue {
-            key: METADATA_KEY.to_string(),
-            value: Some(base64::encode(&metadata_bytes)),
-        }]))
-        .set_compression(Compression::ZSTD)
-        .build()
-}
-
-#[derive(Debug, Default, Clone)]
-struct MemWriter {
-    mem: Arc<Mutex<Cursor<Vec<u8>>>>,
-}
-
-impl MemWriter {
-    /// Returns the inner buffer as long as there are no other references to the
-    /// Arc.
-    pub fn into_inner(self) -> Option<Vec<u8>> {
-        Arc::try_unwrap(self.mem)
-            .ok()
-            .map(|mutex| mutex.into_inner().into_inner())
-    }
-}
-
-impl Write for MemWriter {
-    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        let mut inner = self.mem.lock();
-        inner.write(buf)
-    }
-
-    fn flush(&mut self) -> std::io::Result<()> {
-        let mut inner = self.mem.lock();
-        inner.flush()
-    }
-}
-
-impl Seek for MemWriter {
-    fn seek(&mut self, pos: SeekFrom) -> std::io::Result<u64> {
-        let mut inner = self.mem.lock();
-        inner.seek(pos)
-    }
-}
-
-impl TryClone for MemWriter {
-    fn try_clone(&self) -> std::io::Result<Self> {
-        Ok(Self {
-            mem: Arc::clone(&self.mem),
-        })
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use futures::{stream, StreamExt, TryStreamExt};
-    use iox_catalog::interface::NamespaceId;
-    use parquet::schema::types::ColumnPath;
+    use iox_catalog::interface::{NamespaceId, PartitionId, SequenceNumber, SequencerId, TableId};
     use query::test::{raw_data, TestChunk};
+    use std::sync::Arc;
+    use time::Time;
+    use uuid::Uuid;
 
     fn now() -> Time {
         Time::from_timestamp(0, 0)
@@ -379,15 +187,5 @@ mod tests {
             path.to_raw(),
             format!("1/2/3/4/{}.parquet", metadata.object_store_id)
         );
-    }
-
-    #[test]
-    fn writer_props_have_compression() {
-        // should be writing with compression
-        let props = writer_props(&[]);
-
-        // arbitrary column name to get default values
-        let col_path: ColumnPath = "default".into();
-        assert_eq!(props.compression(&col_path), Compression::ZSTD);
     }
 }
